Checkpoint on removal of global events subsystem. (WIP)
This decision was taken to mitigate the rising complexity of N^2 for N
being all interacting parties over the event system.

Current design aims to provide facades towards each of the main modules.
Gui, Backend, Download etc. Each driving each other through those
defined interfaces.

Within each module an event system may be used, in a wholly encapsulated
way. This is usefull for instance for handling detached state changes
between the requests of the backend towards its server and the responses
coming back in.
It is to be expected, that the server also pushes state
changes without previous request, hence the asymetry and usefulness of
such an event system, or event-reactive approach.